:sparkles: updateProject - suppression des associations companies et participants

+++ src/api/main/java/isen/projet_dp_api/service/ProjectService.java
@@ -50,17 +50,12 @@ public class ProjectService {
         this.companyServiceDAO = companyServiceDAO;
     }
 
-    public void createProjectCompanies(String name, ProjectDAO projectDAO) {
-        projectDAO.addCompagnie(companyServiceDAO.getCompanyByName(name));
-    }
-
     @Transactional
     public ProjectCreationRequestResponse createProject(ProjectDTO projectDTO, String email) {
-
-        var projectDAO = new ProjectDAO(projectDTO, email);
+        var user = userServiceDAO.getUserByEmail(email);
+        var projectDAO = new ProjectDAO(projectDTO, user);
 
         projectDAO = projectServiceDAO.createProject(projectDAO);
-
         var emailError = prepareSendCreationProjectEmail(projectDAO);
         var responseDetails = emailError.map(error -> Map.of(
                 ApiResponseMessage.PROJECT_CREATION, ApiResponseMessage.FAILURE,
@@ -72,17 +67,24 @@ public class ProjectService {
         ));
 
         projectDAO.addParticipant(userServiceDAO.getUserByEmail(email));
-        var projectParticipants = new ArrayList<ParticipantDTO>();
-        var p_chef = new ParticipantDTO();
-        p_chef.setEmail(email);
-        projectParticipants.add(p_chef);
-        projectDTO.setParticipants(projectParticipants);
-
         if (projectDTO.getParticipants() != null) {
             for (var participantDTO : projectDTO.getParticipants()) {
                 projectDAO.addParticipant(userServiceDAO.getUserByEmail(participantDTO.getEmail()));
             }
         }
+
+
+        var projectParticipants = projectDTO.getParticipants();
+
+        if (projectParticipants == null) {
+            projectParticipants = new ArrayList<>();
+        }
+
+        var p_chef = new ParticipantDTO();
+        p_chef.setEmail(email);
+        projectParticipants.add(p_chef);
+        projectDTO.setParticipants(projectParticipants);
+
 
         if (projectDTO.getCompagnies() != null) {
             for (var compagnieDTO : projectDTO.getCompagnies()) {
@@ -90,6 +92,7 @@ public class ProjectService {
 
             }
         }
+        log.info("FIN {}", projectDAO);
 
         var status = emailError.isPresent() ? ApiResponseMessage.PARTIAL_SUCCESS : ApiResponseMessage.SUCCESS;
         var message = ApiResponseMessage.PROJECT_CREATION_SUCCESS + (emailError.isPresent() ? ApiResponseMessage.EMAIL_SEND_ERROR : ApiResponseMessage.EMAIL_SEND_SUCCESS);
@@ -100,8 +103,8 @@ public class ProjectService {
 
     @Transactional
     public ProjectDTO updateProject(ProjectDTO projectDTO, String email, String title) {
-        var project = projectServiceDAO.getProjectByEmailAndTitle(email, title);
 
+        var project = projectServiceDAO.getProjectByEmailAndTitle(email, title);
         if (projectDTO.getTitle() != null && !projectDTO.getTitle().isEmpty()) {
             project.setTitle(projectDTO.getTitle());
         }
@@ -127,14 +130,26 @@ public class ProjectService {
             project.setEnd_date(projectDTO.getEnd_date());
         }
 
+        var projectParticipants = projectDTO.getParticipants();
+
+        if (projectParticipants == null) {
+            projectParticipants = new ArrayList<>();
+        }
+
+        var p_chef = new ParticipantDTO();
+        p_chef.setEmail(email);
+        projectParticipants.add(p_chef);
+        projectDTO.setParticipants(projectParticipants);
 
         if (projectDTO.getParticipants() != null) {
             var currentParticipants = project.getParticipants();
+            log.info("FIN {}", currentParticipants);
+
             var newParticipantsEmails = projectDTO.getParticipants().stream().map(ParticipantDTO::getEmail).toList();
 
             currentParticipants.removeIf(participant -> {
                 if (!newParticipantsEmails.contains(participant.getUser().getEmail())) {
-                    participantServiceDAO.deleteParticipant(participant);
+                    participantServiceDAO.deleteParticipantByUserEmailAndProjectId(participant.getUser().getEmail(), participant.getProject().getId());
                     return true;
                 }
                 return false;
@@ -150,27 +165,26 @@ public class ProjectService {
             }
         }
 
-        var currentCompanies = project.getCompanies();
-        var newCompanyNames = projectDTO.getCompagnies().stream().map(ProjectCompaniesDTO::getName).toList();
 
-        currentCompanies.removeIf(company -> {
-            if (!newCompanyNames.contains(company.getCompany().getName())) {
-                projectCompaniesServiceDAO.deleteProjectCompaniesByCompanyNameAndProjectId(
-                        company.getCompany().getName(), project.getId()
-                );
-                return true;
-            }
-            return false;
-        });
+        if (projectDTO.getCompagnies() != null) {
+            var currentCompanies = project.getCompanies();
+            var newCompanyNames = projectDTO.getCompagnies().stream().map(ProjectCompaniesDTO::getName).toList();
 
-        for (var companyDTO : projectDTO.getCompagnies()) {
-            boolean exists = currentCompanies.stream()
-                    .anyMatch(pc -> pc.getCompany().getName().equals(companyDTO.getName()));
+            currentCompanies.removeIf(company -> {
+                if (!newCompanyNames.contains(company.getCompany().getName())) {
+                    projectCompaniesServiceDAO.deleteProjectCompaniesByCompanyNameAndProjectId(
+                            company.getCompany().getName(), project.getId()
+                    );
+                    return true;
+                }
+                return false;
+            });
 
-            var existingCompany = projectCompaniesServiceDAO.getProjectCompaniesByCompanyNameAndProjectId(companyDTO.getName(), project.getId());
-            log.info(existingCompany);
-            if (existingCompany.isEmpty()) {
-                project.addCompagnie(companyServiceDAO.getCompanyByName(companyDTO.getName()));
+            for (var companyDTO : projectDTO.getCompagnies()) {
+                var existingCompany = projectCompaniesServiceDAO.getProjectCompaniesByCompanyNameAndProjectId(companyDTO.getName(), project.getId());
+                if (existingCompany.isEmpty()) {
+                    project.addCompagnie(companyServiceDAO.getCompanyByName(companyDTO.getName()));
+                }
             }
         }
         projectServiceDAO.updateProject(project);
@@ -212,7 +226,6 @@ public class ProjectService {
         for (var projectCompaniesDAO : companiesDAO) {
             projectCompaniesDTO.add(new ProjectCompaniesDTO(projectCompaniesDAO.getCompany().getName()));
         }
-
         return new ProjectDTO(projectDAO.getEmail().getEmail(), projectDAO.getDescription(), projectDAO.getTitle(), projectDAO.getStatus(),  projectDAO.getStart_date(), projectDAO.getEnd_date(), participantDTO, projectCompaniesDTO);
 
     }
